+++ lib/packages/core/src/api/factories/account/index.ts
@@ -9,3 +9,4 @@ export * from './getAliases';
 export * from './getUnconfirmedAccountTransactions';
 export * from './setAccountInfo';
 export * from './setAlias';
+export * from './setRewardRecipient';