+++ lib/packages/core/src/typings/api/accountApi.ts
@@ -154,4 +154,27 @@ export interface AccountApi {
         senderPrivateKey: string,
         deadline?: number,
     ) => Promise<TransactionId>;
+
+    /**
+     * Assigns a reward recipient for an account
+     *
+     * The transaction will be broadcasted in two steps.
+     * 1. Send the setRewardRecipient call with public key to the network
+     * 2. Take the returned unsigned message and sign it, i.e. the private key won't be transmitted.
+     *
+     * @param recipient The address of the intended reward assignment
+     * @param feeNQT The fee to pay
+     * @param name The name of the account
+     * @param senderPublicKey The senders public key for sending an _unsigned_ message
+     * @param senderPrivateKey The senders private key to _sign_ the message
+     * @param deadline The deadline, in minutes, for the transaction to be confirmed
+     * @return The Transaction ID
+     */
+    setRewardRecipient: (
+        recipient: string,
+        feeNQT: string,
+        senderPublicKey: string,
+        senderPrivateKey: string,
+        deadline?: number,
+    ) => Promise<TransactionId>;
 }