+++ web/angular-wallet/src/app/main/set-reward-recipient/index.ts
@@ -1,0 +1,1 @@
+export { SetRewardRecipientComponent } from './set-reward-recipient.component';